requirement.txt
numpy>=1.20.0
matplotlib>=3.5.0
scipy>=1.7.0
numba>=0.56.0
faiss-cpu>=1.7.0
sentence-transformers>=2.2.0